+++ client/src/navigation/ProfileNavigator.js
@@ -3,6 +3,8 @@ import { createNativeStackNavigator } from "@react-navigation/native-stack";
 import LogInScreen from "../screens/LogInScreen";
 import ProfileScreen from "../screens/ProfileScreen";
 
+import SignInScreen from "../screens/SignInScreen";
+
 const Stack = createNativeStackNavigator();
 
 export default function ProfileNavigator() {
@@ -23,6 +25,15 @@ export default function ProfileNavigator() {
           headerTransparent: true,
         }}
       />
+      <Stack.Screen
+        name="SignIn"
+        component={SignInScreen}
+        options={{
+          title: "",
+          headerTransparent: true,
+        }}
+      />
+
     </Stack.Navigator>
   );
 }
