+++ client/src/screens/LoginScreen.js
@@ -1,75 +1,79 @@
-import React, {useState} from 'react';
+import React, { useEffect, useState } from 'react';
 
 import LogoPintuki from "../components/LogoPintuki.js";
 
-import { Box, Text, Heading, VStack, FormControl, Input, Link, Button, HStack, Center, NativeBaseProvider, Image } from "native-base";
+import { Box, Text, Heading, VStack, Input, Link, Button, HStack, Center, NativeBaseProvider} from "native-base";
+import { getUsuario } from '../services/UsuarioService.js';
 
-import { addUsuario } from '../services/UsuarioService.js';
 
 
-    export default function LogInScreen ({navigation}){
-        
+
+
+    export default function LogInScreen ({navigation, userLogueado , setUserLogueado}){
+      
       const [usuario, setUsuario] = useState('');
       const [contrasenia, setContrasenia] = useState('');
       
-      const registrarUsuario = async () => {
+      const loginUsuario = async () => {
         //el id tiene que ser determinado en el back y el usuario tiene que sacarse el harcodeo
-        let usuarioCreado = {
-          id: 3,
-          usuario: usuario,
-          contrasenia: contrasenia,
-          fotoPerfil: "https://i.ibb.co/KjFFfmq/diego-pintuki-01.jpg",
+        let clave = {
+          username: usuario,
+          password: contrasenia
         };
-        await addUsuario(usuarioCreado)
+        //context provider
+        await setUserLogueado (await getUsuario(clave))
+        console.log("Front", userLogueado)
       };
-      
-      
-      
-      return (
-          <NativeBaseProvider>
-            <Center flex={1} px="3" w="100%">
-              <Box safeArea p="2" py="8" w="90%" maxW="290" >
 
-            <Center>
-                <LogoPintuki/>
+return(
+  <NativeBaseProvider>
+  <Center flex={1} px="3" w="100%">
+    <Box safeArea p="2" py="8" w="90%" maxW="290" >
 
-                <Heading size="lg" fontWeight="600" color="coolGray.800" _dark={{ color: "warmGray.50" }}>
-                Registrarte
-                </Heading>
-            </Center>
-
-        <VStack space={3} mt="5">
-          <FormControl>
-            <FormControl.Label>Nombre de Usuario</FormControl.Label>
-            <Input type="text"
-                  onChangeText={setUsuario}
-                  value={usuario} 
-                  placeholder="Nombre de Usuario"
-                />
-          </FormControl>
-          <FormControl>
-            <FormControl.Label>Constraseña</FormControl.Label>
-            <Input type="password" 
-                  onChangeText={setContrasenia}
-                  value={contrasenia}
-                  placeholder="Constraseña"
-                />
-          </FormControl>
-          {/* <FormControl>
-            <FormControl.Label>Ingresa tu constraseña nuevamente</FormControl.Label>
-            <Input type="password" 
-                  onChangeText={setContrasenia}
-                  value={contrasenia}
-                  placeholder="Constraseña"
-                />
-          </FormControl> */}
-          <Button mt="2" colorScheme="indigo" onPress={registrarUsuario}>
-            Registrarse
-          </Button>
-         
-        </VStack>
-      </Box>
+    <Center>
+      <LogoPintuki/>
+      <Heading size="lg" fontWeight="600" color="coolGray.800" _dark={{ color: "warmGray.50" }}>
+        Iniciar Sesión
+      </Heading>
     </Center>
-          </NativeBaseProvider>
-        );
-    };
+
+    <VStack space={3} mt="5">
+      
+      <Text>Nombre de Usuario</Text>
+      <Input 
+        type="text"
+        onChangeText={setUsuario}
+        value={usuario} 
+        placeholder="Nombre de Usuario"
+      />
+      <Text>Constraseña</Text>
+      <Input 
+        type="password" 
+        onChangeText={setContrasenia}
+        value={contrasenia}
+        placeholder="Constraseña"
+      />
+      
+      <Button mt="2" colorScheme="indigo" onPress={loginUsuario}>
+        Iniciar Sesión
+      </Button>
+      <HStack mt="6" justifyContent="center">
+        <Text fontSize="sm" color="coolGray.600" _dark={{
+        color: "warmGray.200"
+      }}>
+          ¿Todavia no tenes cuenta?.{" "}
+        </Text>
+        <Link _text={{
+        color: "indigo.500",
+        fontWeight: "medium",
+        fontSize: "sm"
+      }} onPress={() => navigation.navigate("SignIn")}>
+          Registrate!
+        </Link>
+      </HStack>
+    </VStack>
+    </Box>
+  </Center>
+  </NativeBaseProvider>
+  );
+}
